+++ pyiron_ontology/__init__.py
@@ -4,7 +4,7 @@ __version__ = get_versions()["version"]
 del get_versions
 
 from pyiron_ontology.atomistics.constructor import AtomisticsOntology
-from pyiron_ontology.tree import build_tree
+from pyiron_ontology.workflow import build_tree
 from pyiron_ontology.dynamic import DynamicOntologies
 
 from pyiron_ontology.atomistics.reasoning import AtomisticsReasoner
